move exceptions to main dependencies
