+++ learnhub_backend/dependencies.py
@@ -1,4 +1,5 @@
 from pydantic import BaseModel
+from fastapi import HTTPException
 
 
 def common_pagination_parameters(skip: int = 0, limit: int = 100):
@@ -7,3 +8,15 @@ def common_pagination_parameters(skip: int = 0, limit: int = 100):
 
 class GenericOKResponse(BaseModel):
     detail: str = "OK"
+
+
+class Exception:
+    bad_request = HTTPException(400, detail="Bad Request")
+    unauthorized = HTTPException(401, detail="Unauthorized")
+    forbidden = HTTPException(
+        403, detail="Forbidden"
+    )  # unlike unauthorized the client's identity is known.
+    not_found = HTTPException(404, detail="Not Found")
+    method_not_allowed = HTTPException(405, detail="Method Not Allowed")
+    request_timeout = HTTPException(408, detail="Request Timeout")
+    teapot = HTTPException(418, detail="I'm a teapot")
